reorganize common and platform specific file
