+++ src/libs/ComposerUtils/updateNumberOfLines/index.js
@@ -1,0 +1,3 @@
+function updateNumberOfLines() {}
+
+export default updateNumberOfLines;
